containerutils/docker: Use WithHost and WithTimeout instead of WithDealContext
This change ensures that when there is an error during the communication
with the Docker Engine API, it prints the socket path it is actually
using.

Before this PR path was always /var/run/docker.sock:
Cannot connect to the Docker daemon at unix:///var/run/docker.sock. Is the docker daemon running?

After this PR it follows that we set in WithHost:
Cannot connect to the Docker daemon at unix:///my/custom/path/. Is the docker daemon running?

Notice that the path comes from client.host:
https://github.com/moby/moby/blob/87a90dc786bda134c9eb02adbae2c6a7342fb7f6/client/errors.go#L22