docs/gadgets: Merge kubectl-gadget and local-gadget instructions
It's very hard for us, and for the users, to have documentation of
gadgets spread around. This commit moves all the gadgets documentation
from local-gadget.md to their respective file in gadgets/.

The current commit doesn't focus on the content, another pass is needed
to check that the content is aligned.