CI: Cache BTFGen
BTFGen takes a lot of time to execute because it has to:
1. clone the btfhub-archive repo (some GBs)
2. uncompress all files from btfhub-archive
3. run btfgen for all btf files uncompressed above

It's only required to run this command when the bpf objects are changed
or new BTF files are pushed to btfhub-archive.

This commit modifies the building system to use the Github actions to
cache the generated BTF files. The changes done here are:
1. Execute btfgen outside the building container: It's hard to extract
  the BTF files from the container to save them in the cache.
2. Define a cache for the btf files: It uses as key the architecture,
  the latest commit of btfhub-archive and the hash of all bpf objects.